Consolidate more methods
