+++ src/GraphQL/Types/Composite/AutoRegisteringObjectGraphType.cs
@@ -44,45 +44,18 @@ namespace GraphQL.Types
             AutoRegisteringHelper.ApplyGraphQLAttributes<TSourceType>(this);
         }
 
-        /// <summary>
-        /// Returns a list of <see cref="FieldType"/> instances representing the fields ready to be
-        /// added to the graph type.
-        /// </summary>
+        /// <inheritdoc cref="AutoRegisteringHelper.ProvideFields(IEnumerable{MemberInfo}, Func{MemberInfo, FieldType?}, bool)"/>
         protected virtual IEnumerable<FieldType> ProvideFields()
-        {
-            foreach (var memberInfo in GetRegisteredMembers())
-            {
-                bool include = true;
-                foreach (var attr in memberInfo.GetGraphQLAttributes())
-                {
-                    include = attr.ShouldInclude(memberInfo, false);
-                    if (!include)
-                        break;
-                }
-                if (!include)
-                    continue;
-                var fieldType = CreateField(memberInfo);
-                if (fieldType != null)
-                    yield return fieldType;
-            }
-        }
+            => AutoRegisteringHelper.ProvideFields(GetRegisteredMembers(), CreateField, false);
 
         /// <summary>
         /// Processes the specified member and returns a <see cref="FieldType"/>.
         /// May return <see langword="null"/> to skip a member.
         /// </summary>
         protected virtual FieldType? CreateField(MemberInfo memberInfo)
-        {
-            var typeInformation = GetTypeInformation(memberInfo);
-            var graphType = typeInformation.ConstructGraphType();
-            var fieldType = AutoRegisteringHelper.CreateField(memberInfo, graphType, false);
-            BuildFieldType(fieldType, memberInfo);
-            // apply field attributes after resolver has been set
-            AutoRegisteringHelper.ApplyFieldAttributes(memberInfo, fieldType, false);
-            return fieldType;
-        }
+            => AutoRegisteringHelper.CreateField(memberInfo, GetTypeInformation, BuildFieldType, false);
 
-        /// <inheritdoc cref="AutoRegisteringHelper.BuildFieldTypeForOutput(MemberInfo, FieldType, Func{MemberInfo, LambdaExpression}, Func{Type, Func{FieldType, ParameterInfo, ArgumentInformation}}, Action{ParameterInfo, QueryArgument})"/>
+        /// <inheritdoc cref="AutoRegisteringOutputHelper.BuildFieldType(MemberInfo, FieldType, Func{MemberInfo, LambdaExpression}, Func{Type, Func{FieldType, ParameterInfo, ArgumentInformation}}, Action{ParameterInfo, QueryArgument})"/>
         protected void BuildFieldType(FieldType fieldType, MemberInfo memberInfo)
         {
             Func<Type, Func<FieldType, ParameterInfo, ArgumentInformation>> getTypedArgumentInfoMethod =
@@ -92,7 +65,7 @@ namespace GraphQL.Types
                     return (Func<FieldType, ParameterInfo, ArgumentInformation>)getArgumentInfoMethodInfo.CreateDelegate(typeof(Func<FieldType, ParameterInfo, ArgumentInformation>), this);
                 };
 
-            AutoRegisteringHelper.BuildFieldTypeForOutput(
+            AutoRegisteringOutputHelper.BuildFieldType(
                 memberInfo,
                 fieldType,
                 BuildMemberInstanceExpression,
@@ -123,75 +96,24 @@ namespace GraphQL.Types
         private ArgumentInformation GetArgumentInformationInternal<TParameterType>(FieldType fieldType, ParameterInfo parameterInfo)
             => GetArgumentInformation<TParameterType>(fieldType, parameterInfo);
 
-        /// <summary>
-        /// Applies <see cref="GraphQLAttribute"/> attributes defined on the supplied <see cref="ParameterInfo"/>
-        /// to the specified <see cref="QueryArgument"/>.
-        /// Also scans the parameter's owning module and assembly for globally-applied attributes.
-        /// </summary>
+        /// <inheritdoc cref="AutoRegisteringOutputHelper.ApplyArgumentAttributes(ParameterInfo, QueryArgument)"/>
         protected virtual void ApplyArgumentAttributes(ParameterInfo parameterInfo, QueryArgument queryArgument)
-        {
-            // Apply derivatives of GraphQLAttribute
-            var attributes = parameterInfo.GetGraphQLAttributes();
-            foreach (var attr in attributes)
-            {
-                attr.Modify(queryArgument);
-            }
-        }
+            => AutoRegisteringOutputHelper.ApplyArgumentAttributes(parameterInfo, queryArgument);
 
-        /// <summary>
-        /// Analyzes a method parameter and returns an instance of <see cref="ArgumentInformation"/>
-        /// containing information necessary to build a <see cref="QueryArgument"/> and <see cref="IFieldResolver"/>.
-        /// Also applies any <see cref="GraphQLAttribute"/> attributes defined on the <see cref="ParameterInfo"/>
-        /// to the returned <see cref="ArgumentInformation"/> instance.
-        /// </summary>
+        /// <inheritdoc cref="AutoRegisteringOutputHelper.GetArgumentInformation{TSourceType}(TypeInformation, FieldType, ParameterInfo)"/>
         protected virtual ArgumentInformation GetArgumentInformation<TParameterType>(FieldType fieldType, ParameterInfo parameterInfo)
-        {
-            var typeInformation = GetTypeInformation(parameterInfo);
-            var argumentInfo = new ArgumentInformation(parameterInfo, typeof(TSourceType), fieldType, typeInformation);
-            argumentInfo.ApplyAttributes();
-            return argumentInfo;
-        }
+            => AutoRegisteringOutputHelper.GetArgumentInformation<TSourceType>(GetTypeInformation(parameterInfo), fieldType, parameterInfo);
 
-        /// <inheritdoc cref="AutoRegisteringHelper.GetRegisteredMembersForOutput{TSourceType}(Expression{Func{TSourceType, object?}}[])"/>
+        /// <inheritdoc cref="AutoRegisteringOutputHelper.GetRegisteredMembers{TSourceType}(Expression{Func{TSourceType, object?}}[])"/>
         protected virtual IEnumerable<MemberInfo> GetRegisteredMembers()
-            => AutoRegisteringHelper.GetRegisteredMembersForOutput(_excludedProperties);
+            => AutoRegisteringOutputHelper.GetRegisteredMembers(_excludedProperties);
 
-        /// <summary>
-        /// Analyzes a property, method or field and returns an instance of <see cref="TypeInformation"/>
-        /// containing information necessary to select a graph type. Nullable reference annotations
-        /// are read, if they exist, as well as the <see cref="RequiredAttribute"/> attribute.
-        /// Then any <see cref="GraphQLAttribute"/> attributes marked on the property are applied.
-        /// <br/><br/>
-        /// Override this method to enforce specific graph types for specific CLR types, or to implement custom
-        /// attributes to change graph type selection behavior.
-        /// </summary>
+        /// <inheritdoc cref="AutoRegisteringHelper.GetTypeInformation(MemberInfo, bool)"/>
         protected virtual TypeInformation GetTypeInformation(MemberInfo memberInfo)
-        {
-            var typeInformation = memberInfo switch
-            {
-                PropertyInfo propertyInfo => new TypeInformation(propertyInfo, false),
-                MethodInfo methodInfo => new TypeInformation(methodInfo),
-                FieldInfo fieldInfo => new TypeInformation(fieldInfo, false),
-                _ => throw new ArgumentOutOfRangeException(nameof(memberInfo), "Only properties, methods and fields are supported."),
-            };
-            typeInformation.ApplyAttributes();
-            return typeInformation;
-        }
+            => AutoRegisteringHelper.GetTypeInformation(memberInfo, false);
 
-        /// <summary>
-        /// Analyzes a method argument and returns an instance of <see cref="TypeInformation"/>
-        /// containing information necessary to select a graph type. Nullable reference annotations
-        /// are read, if they exist, as well as the <see cref="RequiredAttribute"/> attribute.
-        /// Then any <see cref="GraphQLAttribute"/> attributes marked on the property are applied.
-        /// <br/><br/>
-        /// Override this method to enforce specific graph types for specific CLR types, or to implement custom
-        /// attributes to change graph type selection behavior.
-        /// </summary>
+        /// <inheritdoc cref="AutoRegisteringHelper.GetTypeInformation(ParameterInfo)"/>
         protected virtual TypeInformation GetTypeInformation(ParameterInfo parameterInfo)
-        {
-            var typeInformation = new TypeInformation(parameterInfo);
-            typeInformation.ApplyAttributes();
-            return typeInformation;
-        }
+            => AutoRegisteringHelper.GetTypeInformation(parameterInfo);
     }
 }
